# TaekwondoFileServer/Test/Test1/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Taekwondo Orchestration API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Quiz": {
      "get": {
        "tags": [
          "Quiz"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Quiz"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuizDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/QuizDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/QuizDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Quiz/{id}": {
      "get": {
        "tags": [
          "Quiz"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Quiz"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuizDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/QuizDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/QuizDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Quiz"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Quiz/details/{id}": {
      "get": {
        "tags": [
          "Quiz"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Quiz/by-bruger/{brugerId}": {
      "get": {
        "tags": [
          "Quiz"
        ],
        "parameters": [
          {
            "name": "brugerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Quiz/by-klub/{klubId}": {
      "get": {
        "tags": [
          "Quiz"
        ],
        "parameters": [
          {
            "name": "klubId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Quiz/by-pensum/{pensumId}": {
      "get": {
        "tags": [
          "Quiz"
        ],
        "parameters": [
          {
            "name": "pensumId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Quiz/including-deleted/{id}": {
      "get": {
        "tags": [
          "Quiz"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Spørgsmål/{id}": {
      "get": {
        "tags": [
          "Spørgsmål"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Spørgsmål"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpørgsmålDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SpørgsmålDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SpørgsmålDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Spørgsmål"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConflictResolutionStatus": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "QuizDTO": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          },
          "conflictStatus": {
            "$ref": "#/components/schemas/ConflictResolutionStatus"
          },
          "status": {
            "$ref": "#/components/schemas/SyncStatus"
          },
          "lastSyncedVersion": {
            "type": "integer",
            "format": "int32"
          },
          "eTag": {
            "type": "string",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "changeHistoryJson": {
            "type": "string",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "quizID": {
            "type": "string",
            "format": "uuid"
          },
          "quizNavn": {
            "type": "string",
            "nullable": true
          },
          "quizBeskrivelse": {
            "type": "string",
            "nullable": true
          },
          "pensumID": {
            "type": "string",
            "format": "uuid"
          },
          "brugerID": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "klubID": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "spørgsmål": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpørgsmålDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SpørgsmålDTO": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          },
          "conflictStatus": {
            "$ref": "#/components/schemas/ConflictResolutionStatus"
          },
          "status": {
            "$ref": "#/components/schemas/SyncStatus"
          },
          "lastSyncedVersion": {
            "type": "integer",
            "format": "int32"
          },
          "eTag": {
            "type": "string",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "changeHistoryJson": {
            "type": "string",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "spørgsmålID": {
            "type": "string",
            "format": "uuid"
          },
          "spørgsmålRækkefølge": {
            "type": "integer",
            "format": "int32"
          },
          "spørgsmålTid": {
            "type": "integer",
            "format": "int32"
          },
          "teoriID": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "teknikID": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "øvelseID": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "teknik": {
            "$ref": "#/components/schemas/TeknikDTO"
          },
          "teori": {
            "$ref": "#/components/schemas/TeoriDTO"
          },
          "øvelse": {
            "$ref": "#/components/schemas/ØvelseDTO"
          },
          "quizID": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "SyncStatus": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
      "TeknikDTO": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          },
          "conflictStatus": {
            "$ref": "#/components/schemas/ConflictResolutionStatus"
          },
          "status": {
            "$ref": "#/components/schemas/SyncStatus"
          },
          "lastSyncedVersion": {
            "type": "integer",
            "format": "int32"
          },
          "eTag": {
            "type": "string",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "changeHistoryJson": {
            "type": "string",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "teknikID": {
            "type": "string",
            "format": "uuid"
          },
          "teknikNavn": {
            "type": "string",
            "nullable": true
          },
          "teknikBeskrivelse": {
            "type": "string",
            "nullable": true
          },
          "teknikBillede": {
            "type": "string",
            "nullable": true
          },
          "teknikVideo": {
            "type": "string",
            "nullable": true
          },
          "teknikLyd": {
            "type": "string",
            "nullable": true
          },
          "pensumID": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "TeoriDTO": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          },
          "conflictStatus": {
            "$ref": "#/components/schemas/ConflictResolutionStatus"
          },
          "status": {
            "$ref": "#/components/schemas/SyncStatus"
          },
          "lastSyncedVersion": {
            "type": "integer",
            "format": "int32"
          },
          "eTag": {
            "type": "string",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "changeHistoryJson": {
            "type": "string",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "teoriID": {
            "type": "string",
            "format": "uuid"
          },
          "teoriNavn": {
            "type": "string",
            "nullable": true
          },
          "teoriBeskrivelse": {
            "type": "string",
            "nullable": true
          },
          "teoriBillede": {
            "type": "string",
            "nullable": true
          },
          "teoriVideo": {
            "type": "string",
            "nullable": true
          },
          "teoriLyd": {
            "type": "string",
            "nullable": true
          },
          "pensumID": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
      "ØvelseDTO": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          },
          "conflictStatus": {
            "$ref": "#/components/schemas/ConflictResolutionStatus"
          },
          "status": {
            "$ref": "#/components/schemas/SyncStatus"
          },
          "lastSyncedVersion": {
            "type": "integer",
            "format": "int32"
          },
          "eTag": {
            "type": "string",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "changeHistoryJson": {
            "type": "string",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "øvelseID": {
            "type": "string",
            "format": "uuid"
          },
          "øvelseNavn": {
            "type": "string",
            "nullable": true
          },
          "øvelseBeskrivelse": {
            "type": "string",
            "nullable": true
          },
          "øvelseBillede": {
            "type": "string",
            "nullable": true
          },
          "øvelseVideo": {
            "type": "string",
            "nullable": true
          },
          "øvelseTid": {
            "type": "integer",
            "format": "int32"
          },
          "øvelseSværhed": {
            "type": "string",
            "nullable": true
          },
          "pensumID": {
            "type": "string",
            "format": "uuid"
          },
          "brugerID": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "klubID": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}